Refactor backend routes into blueprints
